Added scripts, eslint and modified readme
