+++ .env.example
@@ -1,3 +1,1 @@
-AWS_ACCESS_KEY_ID=<KEY>
-AWS_SECRET_ACCESS_KEY=K<SECRET>
-CONNECTION_TABLE=ConnectionsTable
+CONNECTIONS_TABLE=ConnectionsTable
