+++ .gitignore
@@ -10,3 +10,5 @@ node_modules
 .env
 .idea
 vscode
+*.zip
+.DS_Store
