--- src/aws_sdk_layer/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "GamblingKingsBackend",
-  "version": "1.0.0",
-  "main": "index.js",
-  "repository": "https://github.com/yummyblabla/GamblingKingsBackend.git",
-  "author": "Vincent Wong <vwong125@my.bcit.ca>",
-  "license": "MIT",
-  "devDependencies": {},
-  "dependencies": {
-    "aws-sdk": "^2.687.0"
-  }
-}
