--- src/types.d.ts
@@ -1,10 +1,0 @@
-import { APIGatewayEvent, APIGatewayEventRequestContext } from 'aws-lambda';
-
-interface RealtimeAPIGatewayEventRequestContext extends APIGatewayEventRequestContext {
-  connectionId: string;
-  connectedAt: number;
-}
-
-export interface WebsocketAPIGatewayEvent extends APIGatewayEvent {
-  requestContext: RealtimeAPIGatewayEventRequestContext;
-}
