+++ src/types.ts
@@ -1,0 +1,11 @@
+// eslint-disable-next-line import/no-unresolved
+import { APIGatewayEvent, APIGatewayEventRequestContext } from 'aws-lambda';
+
+interface RealtimeAPIGatewayEventRequestContext extends APIGatewayEventRequestContext {
+  connectionId: string;
+  connectedAt: number;
+}
+
+export interface WebsocketAPIGatewayEvent extends APIGatewayEvent {
+  requestContext: RealtimeAPIGatewayEventRequestContext;
+}
